% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/common.R
\name{filter_overimputed}
\alias{filter_overimputed}
\alias{filter_overimputed,SummarizedExperiment-method}
\alias{filter_overimputed,ExpressionSet-method}
\title{Filtering overimputation of proteomics datasets}
\usage{
filter_overimputed(
  x,
  set.c,
  genes.vc = "all",
  sex.vc = "all",
  imputed_thresh.n = 0.2
)

\S4method{filter_overimputed}{SummarizedExperiment}(
  x,
  set.c,
  genes.vc = "LAT",
  sex.vc = c("M", "F"),
  imputed_thresh.n = 0.2
)

\S4method{filter_overimputed}{ExpressionSet}(
  x,
  set.c,
  genes.vc = "LAT",
  sex.vc = c("M", "F"),
  imputed_thresh.n = 0.2
)
}
\arguments{
\item{x}{An S4 object of class \code{SummarizedExperiment} or \code{ExpressionSet}}

\item{set.c}{Character: name of the data set}

\item{genes.vc}{Character vector: with elements in 'LAT', 'MX2', and 'WT'}

\item{sex.vc}{Character vector: with elements in 'M' and 'F'}

\item{imputed_thresh.n}{Numeric: for proteomics datasets, the features with a
too high proportion of imputed values in all conditions to be compared will be
discarded}
}
\value{
logical vector corresponding to the selected features
}
\description{
Filtering overimputation of proteomics datasets
}
\examples{
set.c <- "proteomics_plasma"

# SummarizedExperiment

proteo.mae <- phenomis::reading(ProMetIS::post_processed_dir.c(),
                                subsets.vc = set.c)
proteo.se <- proteo.mae[[set.c]]
## adding (common) sample metadata to the summarized experiment
sample_meta.DF <- SummarizedExperiment::colData(proteo.mae)
SummarizedExperiment::colData(proteo.se)[, "gene"] <- sample_meta.DF[colnames(proteo.se), "gene"]
SummarizedExperiment::colData(proteo.se)[, "sex"] <- sample_meta.DF[colnames(proteo.se), "sex"]
feat_select.vl <- ProMetIS:::filter_overimputed(proteo.se, set.c = set.c, genes.vc = "LAT")
message(round(sum(feat_select.vl) / length(feat_select.vl) * 100), "\% of selected features")

# ExpressionSet

proteo.mset <- phenomis::reading(ProMetIS::post_processed_dir.c(),
                                 subsets.vc = set.c,
                                 output.c = "set")
proteo.eset <- proteo.mset[[set.c]]
feat_select.vl <- ProMetIS:::filter_overimputed(proteo.eset, set.c = set.c, genes.vc = "LAT")
}
