% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R, R/common.R
\name{imputation_info}
\alias{imputation_info}
\alias{imputation_info,SummarizedExperiment-method}
\alias{imputation_info,ExpressionSet-method}
\title{Checking overimputation in proteomics datasets}
\usage{
imputation_info(x, set.c)

\S4method{imputation_info}{SummarizedExperiment}(x, set.c)

\S4method{imputation_info}{ExpressionSet}(x, set.c)
}
\arguments{
\item{x}{An S4 object of class \code{SummarizedExperiment} or \code{ExpressionSet}}

\item{set.c}{Character: name of the data set}
}
\value{
integer matrix indicating the imputed values in the intensity matrix
}
\description{
Checking overimputation in proteomics datasets
}
\examples{
set.c <- "proteomics_plasma"

# SummarizedExperiment

proteo.mae <- phenomis::reading(ProMetIS::post_processed_dir.c(),
                                subsets.vc = set.c)
proteo.se <- proteo.mae[[set.c]]
imputed.mi <- ProMetIS:::imputation_info(proteo.se, set.c = set.c)
# computing the number of imputed values per feature and per genotype
genotype.fc <- factor(substr(colnames(imputed.mi), 1, 1), levels = c("L", "X", "W"))
geno_impute.mn <- t(apply(imputed.mi, 1, function(feat.vn) tapply(feat.vn, genotype.fc, sum)))
colnames(geno_impute.mn) <- c("LAT", "MX2", "WT")
colSums(geno_impute.mn)
geno_impute.df <- tidyr::gather(as.data.frame(geno_impute.mn), genotype, imputation, LAT:WT, factor_key = TRUE)
ggplot2::ggplot(geno_impute.df, ggplot2::aes(x = imputation, colour = genotype)) + ggplot2::geom_density()

# ExpressionSet

proteo.mset <- phenomis::reading(ProMetIS::post_processed_dir.c(),
                                 subsets.vc = set.c,
                                 output.c = "set")
proteo.eset <- proteo.mset[[set.c]]
imputed.mi <- ProMetIS:::imputation_info(proteo.eset, set.c = set.c)
}
