% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{imputation_metrics}
\alias{imputation_metrics}
\title{Computing the 'imputation' sample nb and percent (i.e. 'imputed by Prostar')
for each variable (for the variables with too high imputation metric in both
genotype conditions to be removed in the statistical analysis )}
\usage{
imputation_metrics(eset)
}
\description{
Computing the 'imputation' sample nb and percent (i.e. 'imputed by Prostar')
for each variable (for the variables with too high imputation metric in both
genotype conditions to be removed in the statistical analysis )
}
