% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{order_mset}
\alias{order_mset}
\title{Moving the selected columns at the beginning of the sampleMetadata and variableMetadata.}
\usage{
order_mset(mset)
}
\description{
Moving the selected columns at the beginning of the sampleMetadata and variableMetadata.
}
