% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProMetIS-package.R
\docType{package}
\name{ProMetIS-package}
\alias{ProMetIS-package}
\alias{ProMetIS}
\title{Multi-omics phenotyping of the LAT and MX2 knockout mice}
\description{
This package provides the data sets and analyzes from the preclinical, proteomics,
and metabolomics study of the LAT and MX2 knockout mice.
}
\examples{
# See the package vignette

}
\author{
E. A. Thévenot (CEA)

Maintainer: Etienne Thevenot <etienne.thevenot@cea.fr>
}
\keyword{package}
